Split App and Webhook fixtures into separate files
