+++ saleor/app/tests/fixtures/__init__.py
@@ -1,0 +1,6 @@
+from .app import *  # noqa: F403
+from .app_extension import *  # noqa: F403
+from .app_installation import *  # noqa: F403
+from .webhooks.payment_app import *  # noqa: F403
+from .webhooks.shipping_app import *  # noqa: F403
+from .webhooks.tax_app import *  # noqa: F403
