+++ saleor/tests/fixtures.py
@@ -6,7 +6,7 @@ from contextlib import contextmanager
 from decimal import Decimal
 from functools import partial
 from io import BytesIO
-from typing import Callable, Optional, Union
+from typing import Callable, Optional
 from unittest.mock import MagicMock
 
 import graphene
@@ -25,8 +25,6 @@ from PIL import Image
 from prices import Money
 
 from ..account.models import Address, Group, StaffNotificationRecipient, User
-from ..app.models import App, AppExtension, AppInstallation
-from ..app.types import AppExtensionMount, AppType
 from ..attribute import AttributeEntityType, AttributeInputType, AttributeType
 from ..attribute.models import (
     Attribute,
@@ -42,7 +40,7 @@ from ..checkout.utils import (
     add_variant_to_checkout,
     add_voucher_to_checkout,
 )
-from ..core import EventDeliveryStatus, JobStatus
+from ..core import JobStatus
 from ..core.models import EventDelivery, EventDeliveryAttempt, EventPayload
 from ..core.payments import PaymentInterface
 from ..core.postgres import FlatConcatSearchVector
@@ -76,7 +74,6 @@ from ..discount.models import (
 )
 from ..giftcard import GiftCardEvents
 from ..giftcard.models import GiftCard, GiftCardEvent, GiftCardTag
-from ..graphql.core.utils import to_global_id_or_none
 from ..menu.models import Menu
 from ..payment import ChargeStatus, TransactionKind
 from ..payment.interface import AddressData, GatewayConfig, GatewayResponse, PaymentData
@@ -124,131 +121,9 @@ from ..warehouse.models import (
     Stock,
     Warehouse,
 )
-from ..webhook.event_types import WebhookEventAsyncType, WebhookEventSyncType
-from ..webhook.models import Webhook, WebhookEvent
-from ..webhook.observability import WebhookData
-from ..webhook.tests.subscription_webhooks import subscription_queries
-from ..webhook.transport.utils import WebhookResponse, to_payment_app_id
+from ..webhook.event_types import WebhookEventAsyncType
+from ..webhook.transport.utils import to_payment_app_id
 from .utils import dummy_editorjs
-
-CALCULATE_TAXES_SUBSCRIPTION_QUERY = """
-subscription CalculateTaxes {
-  event {
-    ...CalculateTaxesEvent
-  }
-}
-
-fragment CalculateTaxesEvent on Event {
-  __typename
-  ... on CalculateTaxes {
-    taxBase {
-      ...TaxBase
-    }
-    recipient {
-      privateMetadata {
-        key
-        value
-      }
-    }
-  }
-}
-
-fragment TaxBase on TaxableObject {
-  pricesEnteredWithTax
-  currency
-  channel {
-    slug
-  }
-  discounts {
-    ...TaxDiscount
-  }
-  address {
-    ...Address
-  }
-  shippingPrice {
-    amount
-  }
-  lines {
-    ...TaxBaseLine
-  }
-  sourceObject {
-    __typename
-    ... on Checkout {
-      avataxEntityCode: metafield(key: "avataxEntityCode")
-      user {
-        ...User
-      }
-    }
-    ... on Order {
-      avataxEntityCode: metafield(key: "avataxEntityCode")
-      user {
-        ...User
-      }
-    }
-  }
-}
-
-fragment TaxDiscount on TaxableObjectDiscount {
-  name
-  amount {
-    amount
-  }
-}
-
-fragment Address on Address {
-  streetAddress1
-  streetAddress2
-  city
-  countryArea
-  postalCode
-  country {
-    code
-  }
-}
-
-fragment TaxBaseLine on TaxableObjectLine {
-  sourceLine {
-    __typename
-    ... on CheckoutLine {
-      id
-      checkoutProductVariant: variant {
-        id
-        product {
-          taxClass {
-            id
-            name
-          }
-        }
-      }
-    }
-    ... on OrderLine {
-      id
-      orderProductVariant: variant {
-        id
-        product {
-          taxClass {
-            id
-            name
-          }
-        }
-      }
-    }
-  }
-  quantity
-  unitPrice {
-    amount
-  }
-  totalPrice {
-    amount
-  }
-}
-
-fragment User on User {
-  id
-  email
-  avataxCustomerCode: metafield(key: "avataxCustomerCode")
-}
-"""
 
 
 class CaptureQueriesContext(BaseCaptureQueriesContext):
@@ -5872,383 +5747,6 @@ def other_description_json():
 
 
 @pytest.fixture
-def app(db):
-    app = App.objects.create(
-        name="Sample app objects",
-        is_active=True,
-        identifier="saleor.app.test",
-        manifest_url="http://localhost:3000/manifest",
-    )
-    return app
-
-
-@pytest.fixture
-def app_marked_to_be_removed(db):
-    app = App.objects.create(
-        name="Sample app objects",
-        is_active=True,
-        identifier="saleor.app.test",
-        manifest_url="http://localhost:3000/manifest",
-        removed_at=timezone.now(),
-    )
-    return app
-
-
-@pytest.fixture
-def webhook_app(
-    db,
-    permission_manage_shipping,
-    permission_manage_gift_card,
-    permission_manage_discounts,
-    permission_manage_menus,
-    permission_manage_products,
-    permission_manage_staff,
-    permission_manage_orders,
-    permission_manage_users,
-):
-    app = App.objects.create(name="Webhook app", is_active=True)
-    app.permissions.add(permission_manage_shipping)
-    app.permissions.add(permission_manage_gift_card)
-    app.permissions.add(permission_manage_discounts)
-    app.permissions.add(permission_manage_menus)
-    app.permissions.add(permission_manage_products)
-    app.permissions.add(permission_manage_staff)
-    app.permissions.add(permission_manage_orders)
-    app.permissions.add(permission_manage_users)
-    return app
-
-
-@pytest.fixture
-def app_with_token(db):
-    app = App.objects.create(name="Sample app objects", is_active=True)
-    app.tokens.create(name="Test")
-    return app
-
-
-@pytest.fixture
-def removed_app(db):
-    app = App.objects.create(
-        name="Deleted app ",
-        is_active=True,
-        removed_at=(timezone.now() - datetime.timedelta(days=1, hours=1)),
-    )
-    return app
-
-
-@pytest.fixture
-def app_with_extensions(app_with_token, permission_manage_products):
-    first_app_extension = AppExtension(
-        app=app_with_token,
-        label="Create product with App",
-        url="www.example.com/app-product",
-        mount=AppExtensionMount.PRODUCT_OVERVIEW_MORE_ACTIONS,
-    )
-    extensions = AppExtension.objects.bulk_create(
-        [
-            first_app_extension,
-            AppExtension(
-                app=app_with_token,
-                label="Update product with App",
-                url="www.example.com/app-product-update",
-                mount=AppExtensionMount.PRODUCT_DETAILS_MORE_ACTIONS,
-            ),
-        ]
-    )
-    first_app_extension.permissions.add(permission_manage_products)
-    return app_with_token, extensions
-
-
-@pytest.fixture
-def removed_app_with_extensions(removed_app, permission_manage_products):
-    first_app_extension = AppExtension(
-        app=removed_app,
-        label="Create product with App",
-        url="www.example.com/app-product",
-        mount=AppExtensionMount.PRODUCT_OVERVIEW_MORE_ACTIONS,
-    )
-    extensions = AppExtension.objects.bulk_create(
-        [
-            first_app_extension,
-            AppExtension(
-                app=removed_app,
-                label="Update product with App",
-                url="www.example.com/app-product-update",
-                mount=AppExtensionMount.PRODUCT_DETAILS_MORE_ACTIONS,
-            ),
-        ]
-    )
-    first_app_extension.permissions.add(permission_manage_products)
-    return removed_app, extensions
-
-
-@pytest.fixture
-def payment_app(db, permission_manage_payments):
-    app = App.objects.create(
-        name="Payment App", is_active=True, identifier="saleor.payment.test.app"
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_payments)
-
-    webhook = Webhook.objects.create(
-        name="payment-webhook-1",
-        app=app,
-        target_url="https://payment-gateway.com/api/",
-    )
-    webhook.events.bulk_create(
-        [
-            WebhookEvent(event_type=event_type, webhook=webhook)
-            for event_type in WebhookEventSyncType.PAYMENT_EVENTS
-        ]
-    )
-    return app
-
-
-@pytest.fixture
-def payment_app_with_subscription_webhooks(db, permission_manage_payments):
-    app = App.objects.create(
-        name="Payment App", is_active=True, identifier="saleor.payment.test.app"
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_payments)
-
-    webhook = Webhook.objects.create(
-        name="payment-subscription-webhook-1",
-        app=app,
-        target_url="https://payment-gateway.com/api/",
-        subscription_query=subscription_queries.PAYMENT_AUTHORIZE,
-    )
-    webhook.events.bulk_create(
-        [
-            WebhookEvent(event_type=event_type, webhook=webhook)
-            for event_type in WebhookEventSyncType.PAYMENT_EVENTS
-        ]
-    )
-    return app
-
-
-@pytest.fixture
-def shipping_app(db, permission_manage_shipping):
-    app = App.objects.create(name="Shipping App", is_active=True)
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_shipping)
-
-    webhook = Webhook.objects.create(
-        name="shipping-webhook-1",
-        app=app,
-        target_url="https://shipping-app.com/api/",
-    )
-    webhook.events.bulk_create(
-        [
-            WebhookEvent(event_type=event_type, webhook=webhook)
-            for event_type in [
-                WebhookEventSyncType.SHIPPING_LIST_METHODS_FOR_CHECKOUT,
-                WebhookEventAsyncType.FULFILLMENT_CREATED,
-            ]
-        ]
-    )
-    return app
-
-
-@pytest.fixture
-def shipping_app_with_subscription(db, permission_manage_shipping):
-    app = App.objects.create(name="Shipping App with subscription", is_active=True)
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_shipping)
-
-    webhook = Webhook.objects.create(
-        name="shipping-webhook-1",
-        app=app,
-        target_url="https://shipping-app.com/api/",
-        subscription_query="""
-        subscription {
-  event {
-    ... on ShippingListMethodsForCheckout {
-      __typename
-    }
-  }
-}
-
-        """,
-    )
-    webhook.events.bulk_create(
-        [
-            WebhookEvent(event_type=event_type, webhook=webhook)
-            for event_type in [
-                WebhookEventSyncType.SHIPPING_LIST_METHODS_FOR_CHECKOUT,
-                WebhookEventAsyncType.FULFILLMENT_CREATED,
-            ]
-        ]
-    )
-    return app
-
-
-@pytest.fixture
-def list_stored_payment_methods_app(db, permission_manage_payments):
-    app = App.objects.create(
-        name="List payment methods app",
-        is_active=True,
-        identifier="saleor.payment.app.list.stored.method",
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_payments)
-
-    webhook = Webhook.objects.create(
-        name="list_stored_payment_methods",
-        app=app,
-        target_url="http://localhost:8000/endpoint/",
-    )
-    webhook.events.create(
-        event_type=WebhookEventSyncType.LIST_STORED_PAYMENT_METHODS,
-    )
-    return app
-
-
-@pytest.fixture
-def stored_payment_method_request_delete_app(db, permission_manage_payments):
-    app = App.objects.create(
-        name="Payment method request delete",
-        is_active=True,
-        identifier="saleor.payment.app.payment.method.request.delete",
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_payments)
-
-    webhook = Webhook.objects.create(
-        name="stored_payment_method_request_delete",
-        app=app,
-        target_url="http://localhost:8000/endpoint/",
-    )
-    webhook.events.create(
-        event_type=WebhookEventSyncType.STORED_PAYMENT_METHOD_DELETE_REQUESTED,
-    )
-    return app
-
-
-@pytest.fixture
-def payment_gateway_initialize_tokenization_app(db, permission_manage_payments):
-    app = App.objects.create(
-        name="Payment method request delete",
-        is_active=True,
-        identifier="saleor.payment.app.payment.gateway.initialize.tokenization",
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_payments)
-
-    webhook = Webhook.objects.create(
-        name="payment_gateway_initialize_tokenization",
-        app=app,
-        target_url="http://localhost:8000/endpoint/",
-    )
-    webhook.events.create(
-        event_type=WebhookEventSyncType.PAYMENT_GATEWAY_INITIALIZE_TOKENIZATION_SESSION,
-    )
-    return app
-
-
-@pytest.fixture
-def payment_method_initialize_tokenization_app(db, permission_manage_payments):
-    app = App.objects.create(
-        name="Payment method initialize tokenization",
-        is_active=True,
-        identifier="saleor.payment.app.payment.method.initialize.tokenization",
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_payments)
-
-    webhook = Webhook.objects.create(
-        name="payment_method_initialize_tokenization",
-        app=app,
-        target_url="http://localhost:8000/endpoint/",
-    )
-    webhook.events.create(
-        event_type=WebhookEventSyncType.PAYMENT_METHOD_INITIALIZE_TOKENIZATION_SESSION,
-    )
-    return app
-
-
-@pytest.fixture
-def payment_method_process_tokenization_app(db, permission_manage_payments):
-    app = App.objects.create(
-        name="Payment method process tokenization",
-        is_active=True,
-        identifier="saleor.payment.app.payment.method.process.tokenization",
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_payments)
-
-    webhook = Webhook.objects.create(
-        name="payment_method_process_tokenization",
-        app=app,
-        target_url="http://localhost:8000/endpoint/",
-    )
-    webhook.events.create(
-        event_type=WebhookEventSyncType.PAYMENT_METHOD_PROCESS_TOKENIZATION_SESSION,
-    )
-    return app
-
-
-@pytest.fixture
-def tax_app(db, permission_handle_taxes):
-    app = App.objects.create(name="Tax App", is_active=True)
-    app.identifier = to_global_id_or_none(app)
-    app.save()
-    app.permissions.add(permission_handle_taxes)
-
-    webhook = Webhook.objects.create(
-        name="tax-webhook-1",
-        app=app,
-        target_url="https://tax-app.com/api/",
-        subscription_query=CALCULATE_TAXES_SUBSCRIPTION_QUERY,
-    )
-    webhook.events.bulk_create(
-        [
-            WebhookEvent(event_type=event_type, webhook=webhook)
-            for event_type in [
-                WebhookEventSyncType.ORDER_CALCULATE_TAXES,
-                WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-            ]
-        ]
-    )
-    return app
-
-
-@pytest.fixture
-def external_tax_app(db, permission_handle_taxes):
-    app = App.objects.create(
-        name="External App",
-        is_active=True,
-        type=AppType.THIRDPARTY,
-        identifier="mirumee.app.simple.tax",
-        about_app="About app text.",
-        data_privacy="Data privacy text.",
-        data_privacy_url="http://www.example.com/privacy/",
-        homepage_url="http://www.example.com/homepage/",
-        support_url="http://www.example.com/support/contact/",
-        configuration_url="http://www.example.com/app-configuration/",
-        app_url="http://www.example.com/app/",
-    )
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_handle_taxes)
-
-    webhook = Webhook.objects.create(
-        name="external-tax-webhook-1",
-        app=app,
-        target_url="https://tax-app.example.com/api/",
-        subscription_query=CALCULATE_TAXES_SUBSCRIPTION_QUERY,
-    )
-    webhook.events.bulk_create(
-        [
-            WebhookEvent(event_type=event_type, webhook=webhook)
-            for event_type in [
-                WebhookEventSyncType.ORDER_CALCULATE_TAXES,
-                WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-            ]
-        ]
-    )
-    return app
-
-
-@pytest.fixture
 def tax_line_data_response():
     return {
         "id": "1234",
@@ -6274,109 +5772,6 @@ def tax_data_response(tax_line_data_response):
         "shipping_tax_rate": 23,
         "lines": [tax_line_data_response] * 5,
     }
-
-
-@pytest.fixture
-def tax_app_with_subscription_webhooks(db, permission_handle_taxes):
-    app = App.objects.create(name="Tax App with subscription", is_active=True)
-    app.permissions.add(permission_handle_taxes)
-
-    webhook = Webhook.objects.create(
-        name="tax-subscription-webhook-1",
-        app=app,
-        target_url="https://tax-app.com/api/",
-        subscription_query=CALCULATE_TAXES_SUBSCRIPTION_QUERY,
-    )
-    webhook.events.bulk_create(
-        [
-            WebhookEvent(event_type=event_type, webhook=webhook)
-            for event_type in [
-                WebhookEventSyncType.ORDER_CALCULATE_TAXES,
-                WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-            ]
-        ]
-    )
-    return app
-
-
-@pytest.fixture
-def observability_webhook(db, permission_manage_observability):
-    app = App.objects.create(name="Observability App", is_active=True)
-    app.tokens.create(name="Default")
-    app.permissions.add(permission_manage_observability)
-
-    webhook = Webhook.objects.create(
-        name="observability-webhook-1",
-        app=app,
-        target_url="https://observability-app.com/api/",
-    )
-    webhook.events.create(event_type=WebhookEventAsyncType.OBSERVABILITY)
-    return webhook
-
-
-@pytest.fixture
-def observability_webhook_data(observability_webhook):
-    return WebhookData(
-        id=observability_webhook.id,
-        saleor_domain="mirumee.com",
-        target_url=observability_webhook.target_url,
-        secret_key=observability_webhook.secret_key,
-    )
-
-
-@pytest.fixture
-def external_app(db):
-    app = App.objects.create(
-        name="External App",
-        is_active=True,
-        type=AppType.THIRDPARTY,
-        identifier="mirumee.app.sample",
-        about_app="About app text.",
-        data_privacy="Data privacy text.",
-        data_privacy_url="http://www.example.com/privacy/",
-        homepage_url="http://www.example.com/homepage/",
-        support_url="http://www.example.com/support/contact/",
-        configuration_url="http://www.example.com/app-configuration/",
-        app_url="http://www.example.com/app/",
-    )
-    app.tokens.create(name="Default")
-    return app
-
-
-@pytest.fixture
-def webhook(app):
-    webhook = Webhook.objects.create(
-        name="Simple webhook", app=app, target_url="http://www.example.com/test"
-    )
-    webhook.events.create(event_type=WebhookEventAsyncType.ORDER_CREATED)
-    return webhook
-
-
-@pytest.fixture
-def webhook_without_name(app):
-    webhook = Webhook.objects.create(app=app, target_url="http://www.example.com/test")
-    webhook.events.create(event_type=WebhookEventAsyncType.ORDER_CREATED)
-    return webhook
-
-
-@pytest.fixture
-def webhook_removed_app(removed_app):
-    webhook = Webhook.objects.create(
-        name="Removed app webhook",
-        app=removed_app,
-        target_url="http://www.example.com/test",
-    )
-    webhook.events.create(event_type=WebhookEventAsyncType.ORDER_CREATED)
-    return webhook
-
-
-@pytest.fixture
-def any_webhook(app):
-    webhook = Webhook.objects.create(
-        name="Any webhook", app=app, target_url="http://www.example.com/any"
-    )
-    webhook.events.create(event_type=WebhookEventAsyncType.ANY)
-    return webhook
 
 
 @pytest.fixture
@@ -6529,15 +5924,6 @@ def checkout_with_prices(
     user.save(update_fields=["metadata"])
 
     return checkout_with_items
-
-
-@pytest.fixture
-def app_installation():
-    app_installation = AppInstallation.objects.create(
-        app_name="External App",
-        manifest_url="http://localhost:3000/manifest",
-    )
-    return app_installation
 
 
 @pytest.fixture
@@ -6794,30 +6180,6 @@ def event_attempt_removed_app(event_delivery_removed_app):
         response="example_response",
         response_headers=None,
         request_headers=None,
-    )
-
-
-@pytest.fixture
-def webhook_response():
-    return WebhookResponse(
-        content="test_content",
-        request_headers={"headers": "test_request"},
-        response_headers={"headers": "test_response"},
-        response_status_code=200,
-        duration=2.0,
-        status=EventDeliveryStatus.SUCCESS,
-    )
-
-
-@pytest.fixture
-def webhook_response_failed():
-    return WebhookResponse(
-        content="example_content_response",
-        request_headers={"headers": "test_request"},
-        response_headers={"headers": "test_response"},
-        response_status_code=500,
-        duration=2.0,
-        status=EventDeliveryStatus.FAILED,
     )
 
 
@@ -7584,326 +6946,6 @@ def setup_mock_for_cache():
         cache_mock.delete = mocked_delete_cache
 
     return _mocked_cache
-
-
-@pytest.fixture
-def setup_checkout_webhooks(
-    permission_handle_taxes,
-    permission_manage_shipping,
-    permission_manage_checkouts,
-):
-    subscription_async_webhooks = """
-    fragment CheckoutFragment on Checkout {
-      shippingPrice {
-        gross {
-          amount
-        }
-      }
-      totalPrice {
-        gross {
-          amount
-        }
-      }
-    }
-
-    subscription {
-      event {
-        ... on CheckoutCreated {
-          checkout {
-            ...CheckoutFragment
-          }
-        }
-        ... on CheckoutUpdated {
-          checkout {
-            ...CheckoutFragment
-          }
-        }
-        ... on CheckoutFullyPaid {
-          checkout {
-            ...CheckoutFragment
-          }
-        }
-        ... on CheckoutMetadataUpdated {
-          checkout {
-            ...CheckoutFragment
-          }
-        }
-      }
-    }
-    """
-
-    def _setup(additional_checkout_event):
-        tax_app, shipping_app, additional_app = App.objects.bulk_create(
-            [
-                App(
-                    name="Sample tax app",
-                    is_active=True,
-                    identifier="saleor.app.tax",
-                ),
-                App(
-                    name="Sample shipping app",
-                    is_active=True,
-                    identifier="saleor.app.shipping",
-                ),
-                App(
-                    name="Sample async webhook app",
-                    is_active=True,
-                    identifier="saleor.app.additional",
-                ),
-            ]
-        )
-        tax_app.permissions.add(permission_handle_taxes)
-        shipping_app.permissions.set(
-            [permission_manage_shipping, permission_manage_checkouts]
-        )
-        additional_app.permissions.add(permission_manage_checkouts)
-        (
-            tax_webhook,
-            shipping_webhook,
-            shipping_filter_webhook,
-            additional_webhook,
-        ) = Webhook.objects.bulk_create(
-            [
-                Webhook(
-                    name="Tax webhook",
-                    app=tax_app,
-                    target_url="http://127.0.0.1/test",
-                    subscription_query="subscription{ event{ ...on CalculateTaxes{ __typename } } }",
-                ),
-                Webhook(
-                    name="Shipping webhook",
-                    app=shipping_app,
-                    target_url="http://127.0.0.1/test",
-                    subscription_query="subscription { event { ... on ShippingListMethodsForCheckout { __typename } } }",
-                ),
-                Webhook(
-                    name="Shipping webhook",
-                    app=shipping_app,
-                    target_url="http://127.0.0.1/test",
-                    subscription_query="subscription { event { ... on CheckoutFilterShippingMethods { __typename } } }",
-                ),
-                Webhook(
-                    name="Checkout additional webhook",
-                    app=additional_app,
-                    target_url="http://127.0.0.1/test",
-                    subscription_query=subscription_async_webhooks,
-                ),
-            ]
-        )
-
-        WebhookEvent.objects.bulk_create(
-            [
-                WebhookEvent(
-                    event_type=WebhookEventSyncType.CHECKOUT_CALCULATE_TAXES,
-                    webhook_id=tax_webhook.id,
-                ),
-                WebhookEvent(
-                    event_type=WebhookEventSyncType.CHECKOUT_FILTER_SHIPPING_METHODS,
-                    webhook_id=shipping_filter_webhook.id,
-                ),
-                WebhookEvent(
-                    event_type=WebhookEventSyncType.SHIPPING_LIST_METHODS_FOR_CHECKOUT,
-                    webhook_id=shipping_webhook.id,
-                ),
-                WebhookEvent(
-                    event_type=additional_checkout_event,
-                    webhook_id=additional_webhook.id,
-                ),
-            ]
-        )
-        return (
-            tax_webhook,
-            shipping_webhook,
-            shipping_filter_webhook,
-            additional_webhook,
-        )
-
-    return _setup
-
-
-@pytest.fixture
-def setup_order_webhooks(
-    permission_handle_taxes,
-    permission_manage_shipping,
-    permission_manage_orders,
-):
-    subscription_async_webhooks = """
-    fragment OrderFragment on Order {
-      shippingPrice {
-        gross {
-          amount
-        }
-      }
-      total {
-        gross {
-          amount
-        }
-      }
-    }
-
-    subscription {
-      event {
-        ... on OrderCreated {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderUpdated {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderPaid {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderExpired {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderRefunded {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderConfirmed {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderFullyPaid {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderFulfilled {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderCancelled {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderBulkCreated {
-          orders {
-            ...OrderFragment
-          }
-        }
-        ... on OrderFullyRefunded {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on OrderMetadataUpdated {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on DraftOrderCreated {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on DraftOrderUpdated {
-          order {
-            ...OrderFragment
-          }
-        }
-        ... on DraftOrderDeleted {
-          order {
-            ...OrderFragment
-          }
-        }
-      }
-    }
-    """
-
-    def _setup(additional_order_event: Union[str, list[str]]):
-        tax_app, shipping_app, additional_app = App.objects.bulk_create(
-            [
-                App(
-                    name="Sample tax app",
-                    is_active=True,
-                    identifier="saleor.app.tax",
-                ),
-                App(
-                    name="Sample shipping app",
-                    is_active=True,
-                    identifier="saleor.app.shipping",
-                ),
-                App(
-                    name="Sample async webhook app",
-                    is_active=True,
-                    identifier="saleor.app.additional",
-                ),
-            ]
-        )
-        tax_app.permissions.add(permission_handle_taxes)
-        shipping_app.permissions.set(
-            [permission_manage_shipping, permission_manage_orders]
-        )
-        additional_app.permissions.add(permission_manage_orders)
-        (
-            tax_webhook,
-            shipping_filter_webhook,
-            additional_webhook,
-        ) = Webhook.objects.bulk_create(
-            [
-                Webhook(
-                    name="Tax webhook",
-                    app=tax_app,
-                    target_url="http://127.0.0.1/test",
-                    subscription_query="subscription{ event{ ...on CalculateTaxes{ __typename } } }",
-                ),
-                Webhook(
-                    name="Shipping webhook",
-                    app=shipping_app,
-                    target_url="http://127.0.0.1/test",
-                    subscription_query="subscription { event { ... on OrderFilterShippingMethods { __typename } } }",
-                ),
-                Webhook(
-                    name="Checkout additional webhook",
-                    app=additional_app,
-                    target_url="http://127.0.0.1/test",
-                    subscription_query=subscription_async_webhooks,
-                ),
-            ]
-        )
-        if isinstance(additional_order_event, str):
-            additional_order_event = [
-                additional_order_event,
-            ]
-        additional_events = [
-            WebhookEvent(
-                event_type=event,
-                webhook_id=additional_webhook.id,
-            )
-            for event in additional_order_event
-        ]
-        WebhookEvent.objects.bulk_create(
-            [
-                WebhookEvent(
-                    event_type=WebhookEventSyncType.ORDER_CALCULATE_TAXES,
-                    webhook_id=tax_webhook.id,
-                ),
-                WebhookEvent(
-                    event_type=WebhookEventSyncType.ORDER_FILTER_SHIPPING_METHODS,
-                    webhook_id=shipping_filter_webhook.id,
-                ),
-            ]
-            + additional_events
-        )
-        return (
-            tax_webhook,
-            shipping_filter_webhook,
-            additional_webhook,
-        )
-
-    return _setup
 
 
 @pytest.fixture
