+++ saleor/webhook/tests/fixtures/__init__.py
@@ -1,0 +1,4 @@
+from .subscription_webhooks import *  # noqa: F403
+from .webhook import *  # noqa: F403
+from .webhook_data import *  # noqa: F403
+from .webhook_response import *  # noqa: F403
